+++ src/egon/data/datasets/chp_etrago.py
@@ -14,7 +14,7 @@ class ChpEtrago(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="ChpEtrago",
-            version="0.0.1",
+            version="0.0.3",
             dependencies=dependencies,
             tasks=(insert),
         )
@@ -41,6 +41,16 @@ def insert():
         DELETE FROM {targets['link']['schema']}.{targets['link']['table']}
         WHERE carrier LIKE '%%CHP%%'
         AND scn_name = 'eGon2035'
+        AND bus0 IN 
+        (SELECT bus_id 
+         FROM {sources['etrago_buses']['schema']}.{sources['etrago_buses']['table']}
+         WHERE scn_name = 'eGon2035'
+         AND country = 'DE')
+        AND bus1 IN 
+        (SELECT bus_id 
+         FROM {sources['etrago_buses']['schema']}.{sources['etrago_buses']['table']}
+         WHERE scn_name = 'eGon2035'
+         AND country = 'DE')
         """
     )
     db.execute_sql(
@@ -88,7 +98,7 @@ def insert():
                     int
                 ),
                 "p_nom": chp_dh.loc[chp_link_dh, "el_capacity"],
-                "carrier": "urban central gas CHP",
+                "carrier": "central_gas_CHP",
             },
         ),
         "eGon2035",
@@ -114,7 +124,7 @@ def insert():
                 "bus0": chp_dh.loc[chp_link_dh, "ch4_bus_id"].astype(int),
                 "bus1": chp_dh.loc[chp_link_dh, "heat_bus_id"].astype(int),
                 "p_nom": chp_dh.loc[chp_link_dh, "th_capacity"],
-                "carrier": "urban central gas CHP heat",
+                "carrier": "central_gas_CHP_heat",
             },
         ),
         "eGon2035",
@@ -141,7 +151,7 @@ def insert():
                 int
             ),
             "p_nom": chp_dh.loc[chp_generator_dh, "el_capacity"],
-            "carrier": "urban central biomass CHP",
+            "carrier": "central_biomass_CHP",
         },
     )
 
@@ -164,7 +174,7 @@ def insert():
             "scn_name": "eGon2035",
             "bus": chp_dh.loc[chp_generator_dh, "heat_bus_id"].astype(int),
             "p_nom": chp_dh.loc[chp_generator_dh, "th_capacity"],
-            "carrier": "urban central biomass CHP heat",
+            "carrier": "central_biomass_CHP_heat",
         },
     )
 
@@ -207,7 +217,7 @@ def insert():
                     chp_link_ind, "electrical_bus_id"
                 ].astype(int),
                 "p_nom": chp_industry.loc[chp_link_ind, "el_capacity"],
-                "carrier": "industrial gas CHP",
+                "carrier": "industrial_gas_CHP",
             },
         ),
         "eGon2035",
@@ -232,7 +242,7 @@ def insert():
                 chp_generator_ind, "electrical_bus_id"
             ].astype(int),
             "p_nom": chp_industry.loc[chp_generator_ind, "el_capacity"],
-            "carrier": "industrial biomass CHP",
+            "carrier": "industrial_biomass_CHP",
         },
     )