+++ src/egon/data/datasets/osmtgmod/__init__.py
@@ -15,6 +15,7 @@ import psycopg2
 from egon.data import db
 from egon.data.config import settings
 from egon.data.datasets import Dataset
+from egon.data.datasets.scenario_parameters import get_sector_parameters
 import egon.data.config
 import egon.data.subprocess as subproc
 
@@ -535,11 +536,15 @@ def to_pypsa():
 
     for scenario_name in ["'eGon2035'", "'eGon100RE'"]:
 
+        capital_cost = get_sector_parameters("electricity", "eGon2035")[
+            "capital_cost"
+        ]
+
         db.execute_sql(
             f"""
             -- BUS DATA
             INSERT INTO grid.egon_etrago_bus (scn_name, bus_id, v_nom,
-                                             geom, x, y, carrier)
+                                             geom, x, y, carrier, country)
             SELECT
               {scenario_name},
               bus_i AS bus_id,
@@ -547,15 +552,17 @@ def to_pypsa():
               geom,
               ST_X(geom) as x,
               ST_Y(geom) as y,
-              'AC' as carrier
+              'AC' as carrier,
+              cntr_id
               FROM osmtgmod_results.bus_data
               WHERE result_id = 1;
 
 
             -- BRANCH DATA
             INSERT INTO grid.egon_etrago_line (scn_name, line_id, bus0,
-                                              bus1, x, r, b, s_nom, cables, v_nom,
-                                              geom, topo)
+                                              bus1, x, r, b, s_nom, s_nom_min, s_nom_extendable,
+                                              cables, v_nom,
+                                              geom, topo, carrier)
             SELECT
               {scenario_name},
               branch_id AS line_id,
@@ -565,10 +572,13 @@ def to_pypsa():
               br_r AS r,
               br_b as b,
               rate_a as s_nom,
+              rate_a as s_nom_min,
+              TRUE,
               cables,
               branch_voltage/1000 as v_nom,
               geom,
-              topo
+              topo,
+              'AC' as carrier
               FROM osmtgmod_results.branch_data
               WHERE result_id = 1 and (link_type = 'line' or
                                        link_type = 'cable');
@@ -577,7 +587,7 @@ def to_pypsa():
             -- TRANSFORMER DATA
             INSERT INTO grid.egon_etrago_transformer (scn_name,
                                                      trafo_id, bus0, bus1, x,
-                                                     s_nom, tap_ratio,
+                                                     s_nom, s_nom_min, s_nom_extendable, tap_ratio,
                                                      phase_shift, geom, topo)
             SELECT
               {scenario_name},
@@ -586,6 +596,8 @@ def to_pypsa():
               t_bus AS bus1,
               br_x/100 AS x,
               rate_a as s_nom,
+              rate_a as s_nom_min,
+              TRUE,
               tap AS tap_ratio,
               shift AS phase_shift,
               geom,
@@ -625,7 +637,62 @@ def to_pypsa():
                  as result
             WHERE a.line_id = result.line_id;
 
+            -- set capital costs for eHV-lines 
+            UPDATE grid.egon_etrago_line
+            SET capital_cost = {capital_cost['ac_ehv_overhead_line']} * length
+            WHERE v_nom > 110;
 
+            -- set capital costs for HV-lines 
+            UPDATE grid.egon_etrago_line
+            SET capital_cost = {capital_cost['ac_hv_overhead_line']} * length
+            WHERE v_nom = 110;
+            
+            -- set capital costs for transformers 
+            UPDATE grid.egon_etrago_transformer a
+            SET capital_cost = {capital_cost['transformer_380_220']}
+            WHERE (a.bus0 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 380)
+            AND a.bus1 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 220))
+            OR (a.bus0 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 220)
+            AND a.bus1 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 380));
+
+            UPDATE grid.egon_etrago_transformer a
+            SET capital_cost = {capital_cost['transformer_380_110']}
+            WHERE (a.bus0 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 380)
+            AND a.bus1 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 110))
+            OR (a.bus0 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 110)
+            AND a.bus1 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 380));
+
+            UPDATE grid.egon_etrago_transformer a
+            SET capital_cost = {capital_cost['transformer_220_110']}
+            WHERE (a.bus0 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 220)
+            AND a.bus1 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 110))
+            OR (a.bus0 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 110)
+            AND a.bus1 IN (
+                SELECT bus_id FROM grid.egon_etrago_bus
+                WHERE v_nom = 220));
+            
             -- delete buses without connection to AC grid and generation or
             -- load assigned
 
@@ -652,7 +719,7 @@ class Osmtgmod(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="Osmtgmod",
-            version="0.0.0",
+            version="0.0.2",
             dependencies=dependencies,
             tasks=(
                 import_osm_data,